+++ render.js
@@ -47,10 +47,10 @@ async function loadSettings() {
 async function updateSoundsList() { 
     let sounds = await ipc.invoke('loadSounds');
     let soundsHTML = `<h3>Sounds</h3>
-    <button class="btn btn-primary btn-sm active" onclick="updateSoundsList()"><svg class="bi bi-arrow-clockwise" width="1em" height="1em" viewBox="0 0 16 16" fill="white" xmlns="http://www.w3.org/2000/svg">
+    <button class="btn btn-primary btn-sm mr-1" onclick="updateSoundsList()"><svg class="bi bi-arrow-clockwise" width="1em" height="1em" viewBox="0 0 16 16" fill="white" xmlns="http://www.w3.org/2000/svg">
     <path fill-rule="evenodd" d="M3.17 6.706a5 5 0 0 1 7.103-3.16.5.5 0 1 0 .454-.892A6 6 0 1 0 13.455 5.5a.5.5 0 0 0-.91.417 5 5 0 1 1-9.375.789z"/>
     <path fill-rule="evenodd" d="M8.147.146a.5.5 0 0 1 .707 0l2.5 2.5a.5.5 0 0 1 0 .708l-2.5 2.5a.5.5 0 1 1-.707-.708L10.293 3 8.147.854a.5.5 0 0 1 0-.708z"/>
-  </svg> Reload Sounds</button><button class="btn btn-primary btn-sm active mx-2" onclick="openSoundsDir()"><svg class="bi bi-folder-symlink" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+  </svg> Reload Sounds</button><button class="btn btn-primary btn-sm mr-1" onclick="openSoundsDir()"><svg class="bi bi-folder-symlink" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
   <path d="M9.828 4a3 3 0 0 1-2.12-.879l-.83-.828A1 1 0 0 0 6.173 2H2.5a1 1 0 0 0-1 .981L1.546 4h-1L.5 3a2 2 0 0 1 2-2h3.672a2 2 0 0 1 1.414.586l.828.828A2 2 0 0 0 9.828 3v1z"/>
   <path fill-rule="evenodd" d="M13.81 4H2.19a1 1 0 0 0-.996 1.09l.637 7a1 1 0 0 0 .995.91h10.348a1 1 0 0 0 .995-.91l.637-7A1 1 0 0 0 13.81 4zM2.19 3A2 2 0 0 0 .198 5.181l.637 7A2 2 0 0 0 2.826 14h10.348a2 2 0 0 0 1.991-1.819l.637-7A2 2 0 0 0 13.81 3H2.19z"/>
   <path d="M8.616 10.24l3.182-1.969a.443.443 0 0 0 0-.742l-3.182-1.97c-.27-.166-.616.036-.616.372V6.7c-.857 0-3.429 0-4 4.8 1.429-2.7 4-2.4 4-2.4v.769c0 .336.346.538.616.371z"/>
@@ -65,7 +65,7 @@ async function updateSoundsList() {
                 buttonRowCount = 0;
                 soundsHTML += `</tr><tr>`;
             }
-            soundsHTML += `<td><button type="button" class="btn btn-secondary" style="height: 100%; width:100%;" onclick="playSound('${sounds[s]}')">${sounds[s]}</button></td>`;
+            soundsHTML += `<td><button type="button" class="btn btn-secondary" style="height: 100%; width:100%;" onclick="playSound('${sounds[s]}')">${sounds[s].replace('.mp3','')}</button></td>`;
             buttonRowCount++;
         }
     }
@@ -164,11 +164,11 @@ async function populateSettings(settingsPage) {
         let result = await ipc.invoke('getSoundsSettings');
         let soundsPageHTML = `<h3>Sounds</h3>`;
         if(result !== undefined) {
-            soundsPageHTML += `<button type="submit" class="btn btn-primary" onclick="saveSoundsForm()">Save</button><button class="btn btn-primary" onclick="openSoundsDir()"><svg class="bi bi-folder-symlink" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+            soundsPageHTML += `<button type="submit" class="btn btn-primary btn-sm mr-1" onclick="saveSoundsForm()">Save</button><button class="btn btn-primary btn-sm mr-1" onclick="openSoundsDir()"><svg class="bi bi-folder-symlink" width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
             <path d="M9.828 4a3 3 0 0 1-2.12-.879l-.83-.828A1 1 0 0 0 6.173 2H2.5a1 1 0 0 0-1 .981L1.546 4h-1L.5 3a2 2 0 0 1 2-2h3.672a2 2 0 0 1 1.414.586l.828.828A2 2 0 0 0 9.828 3v1z"/>
             <path fill-rule="evenodd" d="M13.81 4H2.19a1 1 0 0 0-.996 1.09l.637 7a1 1 0 0 0 .995.91h10.348a1 1 0 0 0 .995-.91l.637-7A1 1 0 0 0 13.81 4zM2.19 3A2 2 0 0 0 .198 5.181l.637 7A2 2 0 0 0 2.826 14h10.348a2 2 0 0 0 1.991-1.819l.637-7A2 2 0 0 0 13.81 3H2.19z"/>
             <path d="M8.616 10.24l3.182-1.969a.443.443 0 0 0 0-.742l-3.182-1.97c-.27-.166-.616.036-.616.372V6.7c-.857 0-3.429 0-4 4.8 1.429-2.7 4-2.4 4-2.4v.769c0 .336.346.538.616.371z"/>
-          </svg> Open sounds folder</button><button class="btn btn-primary" onclick="externalLink('manageRewards')"> Manage Channel Rewards</button><form id="soundsForm"><table class="table table-striped table-hover"><thead><tr><th scope="col">Filename</th><th scope="col">Reward Name (leave unchecked for random)</th></tr></thead><tbody>`;
+          </svg> Open sounds folder</button><button class="btn btn-primary btn-sm mr-1" onclick="externalLink('manageRewards')"> Manage Channel Rewards</button><form id="soundsForm"><table class="table table-striped table-hover"><thead><tr><th scope="col">Filename</th><th scope="col">Reward Name (leave unchecked for random)</th></tr></thead><tbody>`;
             let randomSounds = result.random; 
             if(Object.keys(result.rewards).length > 0) {
                 for(let sound in result.rewards) {
@@ -186,7 +186,7 @@ async function populateSettings(settingsPage) {
                     <input type="text" class="form-control"></div></td></tr>`;
                 }
             }
-            soundsPageHTML += `</tbody></table></form><button type="submit" class="btn btn-primary" onclick="saveSoundsForm()">Save</button>`;
+            soundsPageHTML += `</tbody></table></form><button type="submit" class="btn btn-primary btn-sm" onclick="saveSoundsForm()">Save</button>`;
         }
         document.getElementById('sounds').innerHTML = soundsPageHTML;
     }
@@ -195,7 +195,7 @@ async function populateSettings(settingsPage) {
         let cmdPageHTML = `<h3>Commands</h3>`;
         if(result !== undefined) {
             if(Object.keys(result)) {                
-                cmdPageHTML += `<button type="submit" class="btn btn-primary" onclick="saveCmdForm()">Save</button><form id="cmdForm"><table class="table table-striped table-hover table-sm" style="table-layout:auto;"><thead><th>Enabled</th><th>Command</th></thead><tbody>`;
+                cmdPageHTML += `<button type="submit" class="btn btn-primary btn-sm" onclick="saveCmdForm()">Save</button><form id="cmdForm"><table class="table table-striped table-hover table-sm" style="table-layout:auto;"><thead><th>Enabled</th><th>Command</th></thead><tbody>`;
                 for(let cmd in result) { 
                     cmdPageHTML += `<tr id="${result[cmd].name}"><td style="width: 5px; text-align: center;">`;
                     if(result[cmd].enabled) {
@@ -206,7 +206,7 @@ async function populateSettings(settingsPage) {
                     }
                     cmdPageHTML += `</td><td>${result[cmd].name}</td></tr>`;
                 }
-                cmdPageHTML += `</tbody></table></form><button type="submit" class="btn btn-primary" onclick="saveCmdForm()">Save</button>`;
+                cmdPageHTML += `</tbody></table></form><button type="submit" class="btn btn-primary btn-sm" onclick="saveCmdForm()">Save</button>`;
                 document.getElementById('cmds').innerHTML = cmdPageHTML;
             }
         }